+++ Editor/EasePreviewAttribute.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace LRT.Easing.Editor
+{
+	public class EasePreviewAttribute : PropertyAttribute { }
+}