+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.1.0a] - 14/11/2024
+## [0.2.0] - 15/11/2024
+### Beautify field drawing
+- Add EasePreview attribute to avoid using custom inspector
+- Add EasePreview drawer to draw the attribute
+- Add dependencie to "lrt.utility" for drawing functionnality
+
+## [0.1.0] - 14/11/2024
 ### First alpha release
 - Add static easing class to handle evaluate method